+++ src/CosmosDB/CosmosDB/ManagedCassandra/NewAzManagedCassandraCluster.cs
@@ -57,7 +57,7 @@ namespace Microsoft.Azure.Commands.CosmosDB
             {
                 clusterResource = CosmosDBManagementClient.CassandraClusters.GetWithHttpMessagesAsync(ResourceGroupName, ClusterName).GetAwaiter().GetResult().Body;
             }
-            catch (CloudException e)
+            catch (ErrorResponseException e)
             {
                 if (e.Response.StatusCode != System.Net.HttpStatusCode.NotFound)
                 {